requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
openai
httpx
python-dotenv
sqlalchemy
psycopg2-binary
python-multipart
